+++ CHANGELOG.md
@@ -8,6 +8,12 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## Version 1.0.2 - Unreleased
 
 
+### Changed
+* Register `pathlib.Path` with `ub.repr2`
+* Can now register global `ub.repr2` extensions `ub.repr2.register`
+* Can now register global `ub.hash_data` extensions `ub.hash_data.register`
+
+
 ## Version 1.0.1 - Released 2022-02-20
 
 ### Fixed